requirements.txt
python-docx
PyPDF2
fpdf
reportlab
Flask
sentence-transformers
